iommu: Remove struct iommu_ops *iommu from arch_setup_dma_ops()
mainline inclusion
from mainline-v6.7-rc6
commit 4720287c7bf76e59d19d4dfbdc3f54eeea6fd46b
category: bugfix
bugzilla: https://github.com/RVCK-Project/rvck/issues/142

--------------------------------

This is not being used to pass ops, it is just a way to tell if an
iommu driver was probed. These days this can be detected directly via
device_iommu_mapped(). Call device_iommu_mapped() in the two places that
need to check it and remove the iommu parameter everywhere.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
Signed-off-by: gaorui <gao.rui@zte.com.cn>