ACPI: utils: Return bool from acpi_evaluate_reference()
mainline inclusion
from mainline-v6.8-rc1
commit 6909e0f322b0527fee9fdc54685e6cad69008713
category: feature
bugzilla: https://github.com/RVCK-Project/rvck-olk/issues/40

--------------------------------

There are only 4 users of acpi_evaluate_reference() and none of them
actually cares about the reason why it fails.  All of them are only
interested in whether or not it is successful, so it can return a bool
value indicating that.

Modify acpi_evaluate_reference() as per the observation above and update
its callers accordingly so as to get rid of useless code and local
variables.

The observable behavior of the kernel is not expected to change after
this modification of the code.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
[Resolve merge conflicts in drivers/acpi/thermal.c]
[Resolve merge conflicts in drivers/acpi/utils.c]
Signed-off-by: haijiao.liu <haijiao.liu@sophgo.com>
Signed-off-by: "Rafael J. Wysocki" <rafael.j.wysocki@intel.com>