trace: add bootargs options for trace and ftrace for hibernate resume
dist inclusion
category: feature
bugzilla:

--------------------------------

For trace and ftrace spent about 1s time in kernel bootup of hibernate
resume,but this is not need in this bootup which was mid-process.

This commit add bootargs options 'notrace' and 'noftrace' when this
bootup is hibernate resume dectected by u-boot. Then pass these to
kernel only in current bootup.  Normal bootup NOT pass these options
so won't have impact on trace and ftrace functions.

Signed-off-by: xianbing Zhu <xianbing.zhu@linux.alibaba.com>
Signed-off-by: Han Gao <gaohan@iscas.ac.cn>