RISC-V: KVM: Allow Smnpm and Ssnpm extensions for guests
mainline inclusion
from Linux 6.12-rc5
commit 1851e7836212c76bebb6944bb1541ddcccbea535
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/79

--------------------------------

The interface for controlling pointer masking in VS-mode is henvcfg.PMM,
which is part of the Ssnpm extension, even though pointer masking in
HS-mode is provided by the Smnpm extension. As a result, emulating Smnpm
in the guest requires (only) Ssnpm on the host.
The guest configures Smnpm through the SBI Firmware Features extension,
which KVM does not yet implement, so currently the ISA extension has no
visible effect on the guest, and thus it cannot be disabled. Ssnpm is
configured using the senvcfg CSR within the guest, so that extension
cannot be hidden from the guest without intercepting writes to the CSR.

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>
Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>
Signed-off-by: yechao-w <wang.yechao255@zte.com.cn>